refactor: Refactor semantic release layer
