+++ src/irx/analysis/analyzer.py
@@ -1,0 +1,32 @@
+"""
+title: Semantic analyzer orchestration.
+summary: >-
+  Compose the specialized semantic visitor mixins around a shared analyzer core
+  so traversal, registration, bindings, and rule logic live in smaller modules.
+"""
+
+from __future__ import annotations
+
+from public import public
+
+from irx.analysis.visitors.base import SemanticAnalyzerCore
+from irx.analysis.visitors.control_flow import ControlFlowVisitorMixin
+from irx.analysis.visitors.declarations import DeclarationVisitorMixin
+from irx.analysis.visitors.expressions import ExpressionVisitorMixin
+from irx.analysis.visitors.imports import ImportVisitorMixin
+
+
+@public
+class SemanticAnalyzer(
+    ImportVisitorMixin,
+    DeclarationVisitorMixin,
+    ExpressionVisitorMixin,
+    ControlFlowVisitorMixin,
+    SemanticAnalyzerCore,
+):
+    """
+    title: Concrete semantic analyzer.
+    summary: >-
+      Walk AST nodes, attach semantic sidecars, and delegate reusable policy to
+      the extracted factories, registries, and binding tables.
+    """
